Auto-formatting to comply with PEP8

See merge request PF-SHE/SHE_PPT!57

+++ SHE_PPT/python/SHE_PPT/table_formats/bfd_moments.py
@@ -16,7 +16,8 @@
 # details.
 #
 # You should have received a copy of the GNU Lesser General Public License along with this library; if not, write to
-# the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+# the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+# Boston, MA 02110-1301 USA
 
 from collections import OrderedDict
 
@@ -29,6 +30,7 @@ from astropy.table import Table
 
 
 logger = getLogger(mv.logger_name)
+
 
 class BFDMomentsTableMeta(object):
     """
@@ -67,7 +69,8 @@ class BFDMomentsTableMeta(object):
         # Store the less-used comments in a dict
         self.comments = OrderedDict(((self.version, None),
                                      (self.format, None),
-                                     (self.extname, "#." + mv.shear_estimates_tag),
+                                     (self.extname,
+                                      "#." + mv.shear_estimates_tag),
                                      (self.model_hash, None),
                                      (self.model_seed, None),
                                      (self.noise_seed, None),
@@ -81,12 +84,13 @@ class BFDMomentsTableMeta(object):
                                      (self.bfd_tmpl_sigma_step, None),
                                      (self.bfd_tmpl_sigma_max, None),
                                      (self.bfd_tmpl_xy_max, None),
-                                     (self.validated, "0: Not tested; 1: Pass; -1: Fail")
+                                     (self.validated,
+                                      "0: Not tested; 1: Pass; -1: Fail")
                                      ))
-
 
         # A list of columns in the desired order
         self.all = list(self.comments.keys())
+
 
 class BFDMomentsTableFormat(object):
     """
@@ -115,8 +119,8 @@ class BFDMomentsTableFormat(object):
         self.fits_dtypes = OrderedDict()
         self.lengths = OrderedDict()
 
-        def set_column_properties(name, is_optional = False, comment = None, dtype = ">f4", fits_dtype = "E",
-                                   length = 1):
+        def set_column_properties(name, is_optional=False, comment=None, dtype=">f4", fits_dtype="E",
+                                  length=1):
 
             assert name not in self.is_optional
 
@@ -130,35 +134,56 @@ class BFDMomentsTableFormat(object):
 
         # Table column labels and properties
 
-        self.ID = set_column_properties("SOURCE_ID", dtype = ">i8", fits_dtype = "K")
+        self.ID = set_column_properties(
+            "SOURCE_ID", dtype=">i8", fits_dtype="K")
 
-        self.flags = set_column_properties("FLAGS", dtype = ">i8", fits_dtype = "K")
-        self.fit_class = set_column_properties("FITCLASS", dtype = ">i2", fits_dtype = "I")
+        self.flags = set_column_properties(
+            "FLAGS", dtype=">i8", fits_dtype="K")
+        self.fit_class = set_column_properties(
+            "FITCLASS", dtype=">i2", fits_dtype="I")
 
-        self.x_world = set_column_properties("X_WORLD_CORR", is_optional = False, comment = "deg")
-        self.y_world = set_column_properties("Y_WORLD_CORR", is_optional = False, comment = "deg")
+        self.x_world = set_column_properties(
+            "X_WORLD_CORR", is_optional=False, comment="deg")
+        self.y_world = set_column_properties(
+            "Y_WORLD_CORR", is_optional=False, comment="deg")
 
-        self.x_world_var = set_column_properties("ERRX2_WORLD_CORR", is_optional = True, comment = "deg**2")
-        self.y_world_var = set_column_properties("ERRY2_WORLD_CORR", is_optional = True, comment = "deg**2")
+        self.x_world_var = set_column_properties(
+            "ERRX2_WORLD_CORR", is_optional=True, comment="deg**2")
+        self.y_world_var = set_column_properties(
+            "ERRY2_WORLD_CORR", is_optional=True, comment="deg**2")
 
         # BFD specific columns
-        self.bfd_moments = set_column_properties( "BFD_MOMENTS", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_deriv_moments_dg1 = set_column_properties( "BFD_DM_DG1", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_deriv_moments_dg2 = set_column_properties( "BFD_DM_DG2", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_deriv_moments_dmu = set_column_properties( "BFD_DM_DMU", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dg1dg1 = set_column_properties( "BFD_D2M_DG1DG1", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dg1dg2 = set_column_properties( "BFD_D2M_DG1DG2", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dg2dg2 = set_column_properties( "BFD_D2M_DG2DG2", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dg1dmu = set_column_properties( "BFD_D2M_DG1DMU", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dg2dmu = set_column_properties( "BFD_D2M_DG2DMU", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_2ndderiv_moments_dmudmu = set_column_properties( "BFD_D2M_DMUDMU", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 7 )
-        self.bfd_template_weight = set_column_properties( "BFD_TMPL_WEIGHT", is_optional = True, dtype = ">F8", fits_dtype = "E" )
-        self.bfd_jsuppress = set_column_properties( "BFD_JSUPPRESS", is_optional = True, dtype = ">F8", fits_dtype = "E" )
-        self.bfd_pqr = set_column_properties( "BFD_PQR", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 6 )
+        self.bfd_moments = set_column_properties(
+            "BFD_MOMENTS", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_deriv_moments_dg1 = set_column_properties(
+            "BFD_DM_DG1", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_deriv_moments_dg2 = set_column_properties(
+            "BFD_DM_DG2", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_deriv_moments_dmu = set_column_properties(
+            "BFD_DM_DMU", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dg1dg1 = set_column_properties(
+            "BFD_D2M_DG1DG1", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dg1dg2 = set_column_properties(
+            "BFD_D2M_DG1DG2", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dg2dg2 = set_column_properties(
+            "BFD_D2M_DG2DG2", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dg1dmu = set_column_properties(
+            "BFD_D2M_DG1DMU", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dg2dmu = set_column_properties(
+            "BFD_D2M_DG2DMU", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_2ndderiv_moments_dmudmu = set_column_properties(
+            "BFD_D2M_DMUDMU", is_optional=True, dtype=">F8", fits_dtype="E", length=7)
+        self.bfd_template_weight = set_column_properties(
+            "BFD_TMPL_WEIGHT", is_optional=True, dtype=">F8", fits_dtype="E")
+        self.bfd_jsuppress = set_column_properties(
+            "BFD_JSUPPRESS", is_optional=True, dtype=">F8", fits_dtype="E")
+        self.bfd_pqr = set_column_properties(
+            "BFD_PQR", is_optional=True, dtype=">F8", fits_dtype="E", length=6)
 
-        self.bfd_cov_even = set_column_properties( "BFD_COV_EVEN", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 15 )
-        self.bfd_cov_odd = set_column_properties( "BFD_COV_ODD", is_optional = True, dtype = ">F8", fits_dtype = "E", length = 3 )
-
+        self.bfd_cov_even = set_column_properties(
+            "BFD_COV_EVEN", is_optional=True, dtype=">F8", fits_dtype="E", length=15)
+        self.bfd_cov_odd = set_column_properties(
+            "BFD_COV_ODD", is_optional=True, dtype=">F8", fits_dtype="E", length=3)
 
         # A list of columns in the desired order
         self.all = list(self.is_optional.keys())
@@ -175,22 +200,23 @@ bfd_moments_table_format = BFDMomentsTableFormat()
 # And a convient alias for it
 tf = bfd_moments_table_format
 
-def make_bfd_moments_table_header(detector_x = 1,
-                                      detector_y = 1,
-                                      model_hash = None,
-                                      model_seed = None,
-                                      noise_seed = None,
-                                      detector = None,
-                                      obs_time = None,
-                                      bfd_nlost = None,
-                                      bfd_wt_n = None,
-                                      bfd_wt_sigma = None,
-                                      bfd_tmpl_snmin = None,
-                                      bfd_tmpl_sigma_xy = None,
-                                      bfd_tmpl_sigma_flux = None,
-                                      bfd_tmpl_sigma_step = None,
-                                      bfd_tmpl_sigma_max = None,
-                                      bfd_tmpl_xy_max = None,):
+
+def make_bfd_moments_table_header(detector_x=1,
+                                  detector_y=1,
+                                  model_hash=None,
+                                  model_seed=None,
+                                  noise_seed=None,
+                                  detector=None,
+                                  obs_time=None,
+                                  bfd_nlost=None,
+                                  bfd_wt_n=None,
+                                  bfd_wt_sigma=None,
+                                  bfd_tmpl_snmin=None,
+                                  bfd_tmpl_sigma_xy=None,
+                                  bfd_tmpl_sigma_flux=None,
+                                  bfd_tmpl_sigma_step=None,
+                                  bfd_tmpl_sigma_max=None,
+                                  bfd_tmpl_xy_max=None,):
     """
         @brief Generate a header for a shear estimates table.
 
@@ -208,7 +234,8 @@ def make_bfd_moments_table_header(detector_x = 1,
     """
 
     if detector is not None:
-        logger.warn("'detector' argument for make_*_table_header is deprecated: Use detector_x and detector_y instead.")
+        logger.warn(
+            "'detector' argument for make_*_table_header is deprecated: Use detector_x and detector_y instead.")
         detector_x = detector % 6
         detector_y = detector // 6
 
@@ -217,7 +244,8 @@ def make_bfd_moments_table_header(detector_x = 1,
     header[tf.m.version] = tf.__version__
     header[tf.m.format] = tf.m.table_format
 
-    header[tf.m.extname] = dtc.get_id_string(detector_x, detector_y) + "." + mv.shear_estimates_tag
+    header[tf.m.extname] = dtc.get_id_string(
+        detector_x, detector_y) + "." + mv.shear_estimates_tag
 
     header[tf.m.model_hash] = model_hash
     header[tf.m.model_seed] = model_seed
@@ -239,24 +267,25 @@ def make_bfd_moments_table_header(detector_x = 1,
 
     return header
 
-def initialise_bfd_moments_table(detections_table = None,
-                                     optional_columns = None,
-                                     detector_x = None,
-                                     detector_y = None,
-                                     model_hash = None,
-                                     model_seed = None,
-                                     noise_seed = None,
-                                     detector = None,
-                                     obs_time = None,
-                                     bfd_nlost = None,
-                                     bfd_wt_n = None,
-                                     bfd_wt_sigma = None,
-                                     bfd_tmpl_snmin = None,
-                                     bfd_tmpl_sigma_xy = None,
-                                     bfd_tmpl_sigma_flux = None,
-                                     bfd_tmpl_sigma_step = None,
-                                     bfd_tmpl_sigma_max = None,
-                                     bfd_tmpl_xy_max = None,):
+
+def initialise_bfd_moments_table(detections_table=None,
+                                 optional_columns=None,
+                                 detector_x=None,
+                                 detector_y=None,
+                                 model_hash=None,
+                                 model_seed=None,
+                                 noise_seed=None,
+                                 detector=None,
+                                 obs_time=None,
+                                 bfd_nlost=None,
+                                 bfd_wt_n=None,
+                                 bfd_wt_sigma=None,
+                                 bfd_tmpl_snmin=None,
+                                 bfd_tmpl_sigma_xy=None,
+                                 bfd_tmpl_sigma_flux=None,
+                                 bfd_tmpl_sigma_step=None,
+                                 bfd_tmpl_sigma_max=None,
+                                 bfd_tmpl_xy_max=None,):
     """
         @brief Initialise a shear estimates table based on a detections table, with the
                desired set of optional columns
@@ -278,7 +307,8 @@ def initialise_bfd_moments_table(detections_table = None,
     if detector is not None:
         detector_x, detector_y = dtc.resolve_detector_xy(detector)
 
-    assert (detections_table is None) or (is_in_format(detections_table, detf, strict = False))
+    assert (detections_table is None) or (
+        is_in_format(detections_table, detf, strict=False))
 
     if optional_columns is None:
         optional_columns = []
@@ -297,11 +327,12 @@ def initialise_bfd_moments_table(detections_table = None,
             init_cols.append([])
             dtypes.append((tf.dtypes[colname], tf.lengths[colname]))
 
-    bfd_moments_table = Table(init_cols, names = names, dtype = dtypes)
+    bfd_moments_table = Table(init_cols, names=names, dtype=dtypes)
 
     if detections_table is not None:
         if detector_x is None or detector_y is None:
-            detector_x, detector_y = dtc.get_detector_xy(detections_table.meta[detf.m.extname])
+            detector_x, detector_y = dtc.get_detector_xy(
+                detections_table.meta[detf.m.extname])
         if model_hash is None:
             model_hash = detections_table.meta[detf.m.model_hash]
         if model_seed is None:
@@ -314,21 +345,21 @@ def initialise_bfd_moments_table(detections_table = None,
     if detector_y is None:
         detector_y = 1
 
-    bfd_moments_table.meta = make_bfd_moments_table_header(detector_x = detector_x,
-                                                                   detector_y = detector_y,
-                                                                   model_hash = model_hash,
-                                                                   model_seed = model_seed,
-                                                                   noise_seed = noise_seed,
-                                                                   obs_time = obs_time,
-                                                                   bfd_nlost = bfd_nlost,
-                                                                   bfd_wt_n = bfd_wt_n,
-                                                                   bfd_wt_sigma = bfd_wt_sigma,
-                                                                   bfd_tmpl_snmin = bfd_tmpl_snmin,
-                                                                   bfd_tmpl_sigma_xy = bfd_tmpl_sigma_xy,
-                                                                   bfd_tmpl_sigma_flux = bfd_tmpl_sigma_flux,
-                                                                   bfd_tmpl_sigma_step = bfd_tmpl_sigma_step,
-                                                                   bfd_tmpl_sigma_max = bfd_tmpl_sigma_max,
-                                                                   bfd_tmpl_xy_max = bfd_tmpl_xy_max,)
+    bfd_moments_table.meta = make_bfd_moments_table_header(detector_x=detector_x,
+                                                           detector_y=detector_y,
+                                                           model_hash=model_hash,
+                                                           model_seed=model_seed,
+                                                           noise_seed=noise_seed,
+                                                           obs_time=obs_time,
+                                                           bfd_nlost=bfd_nlost,
+                                                           bfd_wt_n=bfd_wt_n,
+                                                           bfd_wt_sigma=bfd_wt_sigma,
+                                                           bfd_tmpl_snmin=bfd_tmpl_snmin,
+                                                           bfd_tmpl_sigma_xy=bfd_tmpl_sigma_xy,
+                                                           bfd_tmpl_sigma_flux=bfd_tmpl_sigma_flux,
+                                                           bfd_tmpl_sigma_step=bfd_tmpl_sigma_step,
+                                                           bfd_tmpl_sigma_max=bfd_tmpl_sigma_max,
+                                                           bfd_tmpl_xy_max=bfd_tmpl_xy_max,)
 
     assert(is_in_format(bfd_moments_table, tf))
